% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrating.R
\name{rrating}
\alias{rrating}
\title{Simulate Likert-like Item Responses}
\usage{
rrating(n, mean = 3, scale.from = 1, scale.to = 5, sd.norm = 1,
  shift.to.mean = "none")
}
\arguments{
\item{n}{number of observations.}

\item{mean}{mean of the normal distribution to be approximated.}

\item{scale.from}{bottom-end of the possible values in scale.}

\item{scale.to}{top-end of the possible values in scale.}

\item{sd.norm}{standard deviation of the normal distribution to be
approximated.}

\item{shift.to.mean}{if set to "none", the mean of observations will
generally not match the mean of the normal to be approximated; to shift the
observations to match it, set to either "quick", "iterative", or
"round-robin" (algorithms discussed in the details)}
}
\value{
A vector of n length.
}
\description{
Generates random distributions of answers to Likert-type items
answered on a numeric (integer) scale.
}
\details{
TODO
}
\examples{
# Simulate 1000 responses for question "From 1 to 5, how much do you like pizza?"
responses = rrating(1000, mean = 4, scale.from = 1, scale.to = 5)
plot(table(responses))
mean(responses)
#
# mean is not 4, because tails of the normal distribution being approximated
# were binned with the boundaries. To keep mean at four, set the shift option:
responses.shifted = rrating(1000, mean = 4, scale.from = 1, scale.to = 5,
shift.to.mean = "quick")
plot(table(responses.shifted))
mean(responses.shifted)



}

